dts: hikey960: Fix overlapping cma areas
The previous commit f023ef2efd54 ("dts: hikey960: Fixup graphics
errors on 4gb boards") was erronius as I had mistakenly
validated with an earlier version of the patch and when I tried
to optimize the memory layout, I incorrectly calculated the
address and then managed to mistakenly not flash the new dtb to
the board when testing. :P

The current code generates a number of errors in cma init, so
this patch fixes it by using the proper offset and in order to
make it fit, reduces the CMA heap size from 128M to 64M, as we
only currently use ~14M of CMA space (2x 1920x1080x4) for the
framebuffers, so the previous reservation was a bit wasteful.

Change-Id: I7d740224cff1e7a8b60712e1b36efae897ccd062
Signed-off-by: John Stultz <john.stultz@linaro.org>